+++ controller_manager/controller_manager/__init__.py
@@ -24,6 +24,7 @@ from .controller_manager_services import (
     switch_controllers,
     unload_controller,
     get_parameter_from_param_file,
+    bcolors,
 )
 
 __all__ = [
@@ -38,4 +39,5 @@ __all__ = [
     "switch_controllers",
     "unload_controller",
     "get_parameter_from_param_file",
+    "bcolors",
 ]
